Pull production code out into own file.
